backend/fastapi/app/services/gemini.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

genai.configure(api_key=os.getenv('GEMINI_API_KEY'))
model = genai.GenerativeModel('gemini-pro')


async def generate(prompt: str) -> str:
    response = model.generate_content(prompt)
    return response.text
