backend/fastapi/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
google-generativeai
